Add files via upload
To create a hypothetical record of supermarket sales from Quickmart supermarket, I determined the information to be captured in the sheet's columns. Using helper columns in another sheet, I used the functions RANDBETWEEN, VLOOKUP, MATCH, INDEX, and ROWS to populate the first sheet. 
To convert the dynamic values in the populated sheet to a table of records, I used the copy and paste function to transfer the values to another sheet, which was ultimately names the sales records.